% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bayes.R
\name{collect_samples_bayes}
\alias{collect_samples_bayes}
\title{Collect All Sample Predictions from Bayesian Results}
\usage{
collect_samples_bayes(cr)
}
\arguments{
\item{cr}{A \code{calibration_result_multiplate} or \code{calibration_result}.}
}
\value{
Data frame, or NULL if no samples.
}
\description{
Collect All Sample Predictions from Bayesian Results
}
