% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_selection.R
\name{compute_loo}
\alias{compute_loo}
\title{Compute LOO-CV for a Fitted Bayesian Model}
\usage{
compute_loo(bayes_fit)
}
\arguments{
\item{bayes_fit}{Output of \code{\link[=fit_bayes_single]{fit_bayes_single()}}.}
}
\value{
A \code{loo} object (from the loo package).
}
\description{
Extracts the log_lik generated quantity and computes PSIS-LOO.
}
