% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bayes.R
\name{stan_model_path}
\alias{stan_model_path}
\title{Get Path to a curveRbayes Stan Model File}
\usage{
stan_model_path(model_family)
}
\arguments{
\item{model_family}{Character. One of the curveRcore model names.}
}
\value{
Absolute path to the \code{.stan} file.
}
\description{
Get Path to a curveRbayes Stan Model File
}
\keyword{internal}
